Change bounded-rdeps visitor to output unique results.
The bounded-rdeps query function(s) are those that specify a `depth` parameter
and return only those reverse-dependencies up to `depth` targets away from
targets in the `<expression>` parameter. For each target we visit during the
function evaluation, we track the minimum-depth at which we visited the target.
If we discover a shorter path to that target, we need to visit it again, as
this might have made the visitation of other targets possible within the maximum
depth.

The `MinDepthUniquifier` is the class used to track this state per-target.
While this class has proven effective at implementing the depth bound, it
cannot be used to ensure that the query function only *outputs* each target
exactly once, not without modification. Bounded-depth query functions have
been producing duplicate results which are then deduplicated by the `Callback`
supplied by the `AbstractBlazeQueryEnvironment`. The `SkyQueryEnvironment` may
be extended such that it does *not* deduplicate in its `Callback` to support
arbitrarily large output sizes, and these environments have been emitting duplicate
results.

This change extends the implementation of the `MinDepthUniquifier` to track both the
min-depth of each target as well as whether that target has been produced as
function output. This ensures the query function itself produces deduplicated
output, which in turn ensures that even extensions of the `SkyQueryEnvironment`
produce deduplicated output.

RELNOTES: n/a
PiperOrigin-RevId: 789563340
Change-Id: Ia8591defa98aa5e2c520bf0db557fee2ae8dc383