[8.4.0] Support generational ZGC and generational Shenandoah (#26684)
Bazel currently does not start when you launch the server JVM with
generational ZGC using `-XX:+UseZGC` and `-XX:+ZGenerational`. Or if
you're on JDK 24+ the same problem is encountered with just
`-XX:+UseZGC`.

The same problem is encountered with `-XX:+UseShenandoahGC` and
`-XX:ShenandoahGCMode=generational`.

Non-generational ZGC is removed as of JDK 24. Generational Shenandoah is
being promoted from experimental to production in JDK 25.

This fix is modeled after #12644

Closes #26676.

PiperOrigin-RevId: 789746919
Change-Id: Ia070fa9b39cef07cc9987544c014d589b2504dce

Commit
https://github.com/bazelbuild/bazel/commit/3a1fe64444ead88b1951f4b7a1eb9b5a9ae6d8fa

Co-authored-by: James Judd <james@lab-y.com>