Keep empty directories in runfiles
Empty directories that are regular inputs were already kept around during spawn input expansion, empty directories in runfiles weren't.

This was found while working on a "golden test" for Merkle tree computation that pins down the Merkle tree hash for a complex spawn. Since this hash is determined purely by the RBE spec, it makes sense to have a test that verifies Bazel's compatibility regardless of its particular RBE implementation. This will be used to catch regressions when changing to a new implementation.

Work towards #21378

Closes #26623.

PiperOrigin-RevId: 792086827
Change-Id: Iab0691582d785030ceac4970067dc709e7b9eb0d