Support packages assembled from package pieces in query and genquery machinery
For a package loaded with lazy macro expansion, Target.getPackage() returns null,
so we now need to use TargetProvider/PackageProvider to retrieve a target's
package's BUILD file target or to iterate sibling targets.

As a follow-up, we'd want to avoid unnecessarily loading full packages in
implementations of TargetProvider.getTarget() and in TargetLoadingUtil -
but that requires another new skyfunction.

Working towards https://github.com/bazelbuild/bazel/issues/23852

PiperOrigin-RevId: 792276714
Change-Id: I76266c10d4f5d6e087a6f38c6b6bfe06eb72603b