Strictify exception types in `remote` package to reduce wrapping
This makes it easier to understand what kinds of exceptions can be thrown by a particular function and avoids wrapping in `RuntimeException` that can turn regular failures into crashes.

In particular, this resolves an issue seen on a PR in which `RemoteSpawnRunner` turned an `ExecException` raised by `RES.uploadInputsIfNotPresent` due to a lost input into a `RuntimeException`, thus resulting in a crash. It's not clear whether this behavior can also happen at HEAD or whether all callers unwrap the `RuntimeException`, so this change doesn't necessarily fix a bug.

Work towards #21378

Closes #26650.

PiperOrigin-RevId: 792329428
Change-Id: Iced921ce8633926e6488d38502c9b93c9fffa393