Fix comparator used in `ConcurrentArtifactPathTrie`
The default comparator for `PathFragment` may sort unrelated paths between a path and a child path, which results in `contains` behaving incorrectly as soon as paths such as `foo/bar` and `foo/bar-baz` are present in the set. This is a bug introduced by #26639.

This is fixed by introducing a new `Comparator` for `PathFragment` that sorts `/` lower than any other character. This will also be used in future work on #21378 and is thus placed in `PathFragment`. It may become the default in the future, but not until Blaze has been confirmed to not rely on the order.

Closes #26714.

PiperOrigin-RevId: 792519419
Change-Id: I8bf8fc6b15cc5bcf0d303057a828fbd644032bab