Include targets with non-idempotent rule transition in query output
Rule transitions that are not idempotent are owned by a `ConfiguredTargetKey` that has `shouldApplyRuleTransition` set to `false`. Since this is not the default, `cquery` and `aquery`'s iteration over all configurations in the build does not pick up these configured targets and their filtering logic removed them when encountered.

This is fixed by augmenting the filtering logic to keep these targets if they wouldn't be encountered in any other way.

Adding a suitable test required removing logic in a test helper that uniquified results.

Closes #26738.

PiperOrigin-RevId: 794272905
Change-Id: Ie6fc0716bf75b77c287ff9e09163aeaee9692584