Clear out stale rdeps before running analysis postprocessors
The deletion of dirty nodes in the Skyframe graph is delayed for performance reasons, which resulted in rdeps pointing to stale nodes being present when `aquery` or `cquery` walked the graph. Fix this by cleaning up these nodes eagerly when running with an analysis postprocessor.

The PR has been manually verified on the reproducer provided in https://github.com/bazelbuild/bazel/issues/15079#issuecomment-1695781700. The reproducer no longer failed at HEAD since 099f0e50610c9abbdb71c6d88553f99df7bead9f, as it apparently required a more complex setup to evade the existing `null` checks and produce a crash.

The test added with this PR is simpler and did not produce an NPE even prior to this change due to the protective `null` checks. It does produce a failure before this PR and passes with it when these checks are patched out:

```diff
--- a/src/main/java/com/google/devtools/build/lib/query2/PostAnalysisQueryEnvironment.java
+++ b/src/main/java/com/google/devtools/build/lib/query2/PostAnalysisQueryEnvironment.java
@@ -399,9 +399,7 @@ public abstract class PostAnalysisQueryEnvironment<T> extends AbstractBlazeQuery
       }
       T rdepValue = getValueFromKey(rdepKey);
       if (rdepValue == null) {
-        // Cannot find the actual value, possibly because it failed during analysis.
-        // TODO: b/324419258 - Add a test for this case
-        continue;
+        throw new RuntimeException("Configured target value for " + rdepKey + " not found.");
       }
       var actualParentKey = getConfiguredTargetKey(rdepValue);
       if (actualParentKey.equals(child)) {
@@ -429,9 +427,7 @@ public abstract class PostAnalysisQueryEnvironment<T> extends AbstractBlazeQuery
       }
       T rdepValue = getValueFromKey(rdepKey);
       if (rdepValue == null) {
-        // Cannot find the actual value, possibly because it failed during analysis.
-        // TODO: b/324419258 - Add a test for this case
-        continue;
+        throw new RuntimeException("Configured target value for " + rdepKey + " not found.");
       }
       var actualParentKey = getConfiguredTargetKey(rdepValue);
       if (!actualParentKey.equals(child)) {
```

Speculatively fixes #26312
Speculatively fixes #23022
Fixes #15079

Closes #26746.

PiperOrigin-RevId: 795100780
Change-Id: I0440612f2a5e8ea6860ef687025eab53fb2ae4ab