Sandboxing: Fix flaky test due to static state
AbstractContainerizingSandboxSpawnTest checked the static boolean in
SandboxHelpers but depending on the other of execution of tests that boolean
may have already been set by the time the flaky test was executed. If it had
already set then no warning would be emitted and the warning counter would be
0.

To fix that the single method that uses the boolean is kept but calls to an
instance method that is called on a static INSTANCE variable. The other methods
are kept static as they were. The test has visibility into the instance method
and calls it directly with a new instance.

RELNOTES:none
PiperOrigin-RevId: 796882185
Change-Id: I690936dd1321011cf5f78284532f6bf1f1b1874b