Fix crash after initial build failure.
This bug was caused by not triggering
SkyframeExecutor.syncRemoteAnalysisCachingState in the case of build
failure. Crucially, this updates the version, which is otherwise left
at null, leading to the failure.

To fix this, splits
BuildTool.logAnalysisCachingStatsAndMaybeUploadFrontier into
serializeValues and logAnalysisCachingStats methods. The latter calls
SkyframeExecutor.syncRemoteAnalysisCachingState and is now always invoked
in the finally clause.

Cleans up the exception handling of logAnalysisCachingStats. It now
handles a possible internal InterruptedException by propagating and
logging. This is necessary because throwing from finally overrides other
thrown exceptions.

PiperOrigin-RevId: 798916995
Change-Id: Ia596e205bf13ac8bb74136c84e9b55260c8d63b6