Fix precondition error in sandboxing
The precondition exists to make sure that we are not littering the sandbox base with untracked directories. This change fixes the scenario where the linux-sandbox crashed when unable to block certain paths due to permissions. The linux-sandbox would not be added to the list of runners after crashing with an IOException and it wouldn't be cleaned up. With this change it fail gracefully.

Work towards fixing #23880. It fixes the second reported error, cannot verify if it fixes the first one too.

RELNOTES:none
PiperOrigin-RevId: 800027300
Change-Id: I20041f477d984372805f7d84af0b0b1606cb46be