Upgrade protobuf to 32.0
Upgrade also collateral modules and MODULE.tools.

Remove rules_proto. The dependency was there because of gRPC, not used directly. But now it's pulled in by protobuf.

Update the protobuf.patch. Most of the diff is gone. New is diff to support gRPC. Upgrading gRPC could possibly help as well.

Needed to Starlarkify CcInfo (which replaces ``.debug_context()`` with ``._debug_context`).

Python errors started to diverge between Bazel and Blaze.

PiperOrigin-RevId: 800030128
Change-Id: I23319093d135c63a44a1ad26f09a06d034874027