Maintain Jacoco's order for branches.
Effectively a redo of commit a915b005, which made incorrect assumptions
about how branch indices were assigned.

This is preparation for an update to Jacoco in which its filters will
become sensitive to branch order for the branches on a line,
particularly for "branch replacement".

This change alters the visitor to record branch indices correctly where
before we were somewhat sloppy (since we didn't use that information).

This will simplify things for us when we update the version of Jacoco we
use.

It's worth noting that "branch indices" do not form a "dense set". An
instruction may have three branches, but be referred to via indices [0,
1, 3]. To account for this the BranchExp object may store a "NullExp"
reference in the unused slots; these will be skipped over when returning
the final coverage data. These BranchExp objects correspond to Jacoco's own
"coveredBranches" BitSet on its Instruction class.

Jacoco's ordering of branches is not necessarily what might be
intuitively expected: given a conditional jump to a target, the "first"
branch is actually the fall-through, with the second branch being the
continuation at the target.
For switch statements, it's the case that the first branch is always the
"default" case.

As a result, there may be an observable change in the order of branches
in the final coverage output files for Java code. However, for branches
that have been "replaced" (in some switch statements for instance), this
will not occur since the new branches will continue to be output in
"walk" order instead of Jacoco's "logical" order. In any case, we have
never made any promises with respect to branch order so this should not
matter.

PiperOrigin-RevId: 800458650
Change-Id: I79d0b39cc7b151cc38f340948e0f154ce431a9fd