Rollforward of: Move `CcInfo` to Starlark
All the substructures of CcInfo are in Starlark already, with the exception of CcCompilationContext.

There are still some test failures. Those are coming from the diffbase and need protobuf release. I'm sending this for an early review.

NEW: Fix handling Nones on CcInfo construction. Nones are replaced with an empty contexts.
PiperOrigin-RevId: 801758070
Change-Id: I54499aac4ebf622109e5e64651327ed3af35ddd7