Starlarkify create_compile_source_action.
We no longer pass around the "compile action builder", instead we
create a fresh one just before calling the "create_cpp_compile_action".
The reasoning is that this builder was often modified or duplicated for
each compilation action anyway and the majority of arguments were
already being passed around so having it created early did not save us
much.

The name of the function is altered to reflect the fact it generates
actions for both PIC and no-PIC compilations. The original name is used
for the "inner" function which generates actions for a given PIC or
no-PIC compile.

_get_compile_output now calls a new internal method to create the output
files. The handling of external repositories is non trivial to implement
in Starlark and there are other subtleties involved, such as never
writing to the genfiles directory.

PiperOrigin-RevId: 801837069
Change-Id: I9cf2d772f4407112b5cb4f5f6c05af848850ca81