[8.4.0] Fix runfiles directory staleness issues caused by symlinking the output manifest (#26868)
`SymlinkTreeAction` is only executed for its side effect of updating the
runfiles tree, but its up-to-dateness is tracked solely by the digest of
its output, the manifest under the runfiles directory. If this manifest
is a symlink to the input manifest and a previous build modified the
runfiles directory while not updating the action cache entry for
`SymlinkTreeAction`, this would cause Bazel to consider the action
up-to-date even though the runfiles directory is stale.

The fix consists of two parts:
* Always copy the manifest instead of symlinking it, matching the
behavior of the former `--noexperimental_inprocess_symlink_creation`
flag. This fixes staleness issues when only using versions of Bazel with
this fix.
* Detect stale state left behind by older versions of Bazel by skipping
action cache hits if the output manifest is a symlink. Since this
situation is no longer created by fixed versions of Bazel, we could
consider removing this logic in the future.

Fixes #26818

Closes #26837.

PiperOrigin-RevId: 801798862
Change-Id: I02c6d1b353981249a98df224efc4f46c5e4866d7

(cherry picked from commit 585e16c701a6eef9a0b96cce31cae56893f04278)

Also includes:
* a3eb3516b26ddaa2aa81d15a3cbfb6876a07d359

Fixes #26824

---------

Co-authored-by: Googler <tjgq@google.com>