Support `ProxyFileArtifactValue` in the action cache.
When the action cache is configured to store remote metadata, also store the exec paths of outputs which are `ProxyFileArtifactValue`. This special type of metadata is used to signal that the source of truth for an output is in the local output base, when an action file system would ordinarily not look for files on disk.

PiperOrigin-RevId: 802196384
Change-Id: Ica7cb075a64c0e34f2ecb635e7606cc7f829a0ff