Support empty tree artifacts when storing output metadata in the action cache.
An empty tree artifact output is common for test actions. The undeclared outputs directory is a tree artifact, but many tests don't have any undeclared outputs.

Mechanically, the support is accomplished by always storing a `SerializableTreeArtifactValue`, even if the tree artifact contains no remote outputs.

Also fixes an issue in `ActionCacheCheckerTest` where `FakeInputMetadataHandler` was returning an empty `TreeArtifactValue` when the root directory did not exist. It should throw `FileNotFoundException` to match the behavior of `ActionOutputMetadataStore`.

PiperOrigin-RevId: 802218210
Change-Id: Ica0cc78ad423a960661a71e3f45b4f54bda2bda2