Add a createSymbolicLink argument indicating the file type of the target.
On Windows, a different kind of filesystem object (junction or symlink) needs to be created depending on the target type, which is not always inspectable in "building without the bytes" scenarios where the symlink is created ahead of its target.

This CL only introduces the formal argument and adjusts the WindowsFileSystem#createSymbolicLink implementation accordingly. A followup CL will pass in the argument where required.

Progress on #21747 and #26701.

PiperOrigin-RevId: 802504976
Change-Id: I912a84ba6734bf0562f369c7659bcc25f259e70f