Do not store in-memory files on disk
The documentation for --experimental_inmemory_dotd_files
and --experimental_inmemory_jdeps_files states
'...instead of being written to disk,' but the files were
actually written to both memory and disk (unless combined
with BwoB). This caused two problems:

1) Unnecessary overhead from remotely downloading all
   in-memory files twice.

2) Paradoxically reduced robustness of
   --remote_download_all: that flag normally increases
   Bazel’s resilience to remote cache issues, but because
   these outputs were kept in memory instead of being
   written to disk by RemoteSpawnCache.lookup (which
   handles cache issues gracefully), action finalization
   treated them as missing on disk and attempted a more
   fragile download, increasing the risk for
   "unable to finalize action" failures (see #22387).

This change ensures that in-memory files are kept
exclusively in memory and are not written to disk. This
aligns with the documentation and addresses both problems.

Fixes #22387