Wait for Skycache write to complete, even if it fails, before returning.
This is achieved by changing the write status future from a
ListenableFuture<Void> to a ListenableFuture<ImmutableList<Throwable>>.
This makes it so that instead of failing fast on error, it waits,
unconditionally, for all the underlying work to complete.

Does some short-circuiting of uploads if an error occurs early.

Also, propagate MissingSkyframeEntryException instead of checkNotNull
in SelectedEntrySerializer.

Remove misleading "Waiting for write futures took an additional"
INFO message. It used to refer only to the time spent waiting for the
writeStatus but this got corrupted somewhere.

PiperOrigin-RevId: 802628482
Change-Id: I3e5382cfbffe826eeae333a5f20a22ff71dd3d8a