Expose global variable doc comments in Starlark programs
Doc comments are static, parser/resolver-level concepts - associated with
variables, not values - so we expose them in Program, not in Module.

Note this means there isn't any reasonable way, unfortunately, to inherit
the doc comments of a value loaded from a different .bzl file. (We could,
hypothetically, inherit doc comments if the rhs of an assignment is the
identifier of a loaded symbol - although this would require *a lot* of
refactoring; but if the rhs is something which evaluates to the value of
a loaded symbol, inheriting docs would require tracking values' provenance,
which is something that we do not want to ever do.)

Working towards https://github.com/bazelbuild/stardoc/issues/202

PiperOrigin-RevId: 802812656
Change-Id: Ia340d94150926aabd7eb9f4665c4d2af15a879b8