Add `Sequence[T]` type to Starlark.
This change introduces `Sequence[T]` as a new type in Starlark's type system. `Sequence[T]` is a supertype of `list[T]` and a subtype of `Collection[T]`. This allows for more precise type annotations for functions that accept any ordered, indexable collection.

Works on: https://github.com/bazelbuild/bazel/issues/25948

PiperOrigin-RevId: 802986800
Change-Id: I6572e19eddcab28957abb2796c21191417d50263