Allow manually specifying the server checksum Skycache uses.
This is useful when one wants to debug an issue with the cache but needs to add debugging logic. Normally, the logic would change the server checksum and thus make it impossible to get cache hits and this mechanism allows one to step around that.

Only allowed in download mode because using data from one server version in another is dangerous enough already, but contaminating the cache would be even worse.

RELNOTES: None.
PiperOrigin-RevId: 803021806
Change-Id: Idc2af7882f703a6504fe646d98544c0c2154a375