Use a primitive array to store line data for code coverage.
Line coverage data is an associated between a line number and an
execution value: an array is a much more compact and efficient way to
store this compared to a HashMap.

Reading a ~800MB file can now be done with a maximum heap of ~4GB
instead of requiring >6GB, with far less thrashing when constrained
to such.

GitHub Issue: https://github.com/bazelbuild/bazel/issues/26383

PiperOrigin-RevId: 803449772
Change-Id: Ifd5fcc4f7f6b7b3d8c219fa3f3b7eb7ca0743aa3