Document a benign memory leak in the client.
A refactor could easily address this but for now I just wanted to acknowledge it for the benefit of future code readers. When reading this code recently, I was surprised to see this. Looking through the code history, it was first introduced in https://github.com/bazelbuild/bazel/commit/4a45d92130a6b1306a3840d006df165b8040a6cf.

PiperOrigin-RevId: 803522746
Change-Id: I84f0073436cc50720213fcbacee124630be81b76