Improve ordering of single-toolchain resolution debugging.
A single-toolchain resolution invocation produces a lot of messages when `--toolchain_resolution_debug` is set. It crucially starts with `Performing resolution of <toolchain type> for target platforn <target platform>`. This is important to understand the context of the other messages that appear in its sequence.

Before this change, `Rejection toolchain <toolchain that doesn't fit>` messages appeared first. This was hard to read because we didn't yet have the context to properly interpret that.

PiperOrigin-RevId: 803565073
Change-Id: I3ae96baa6d8d2685b51f09d2b6d6b77a902d20fa