Speculative fix for repo contents cache race
The error message observed in #26713 is consistent with `FileChannel#open` failing because the given path doesn't exist. This could happen if one Bazel process observed that `!entryDir.isDirectory()`, which is true if the path doesn't exist, and proceeded to delete the directory while another was between creating the directory and opening the channel. Since the entry dir is never expected to be an existing non-directory unless the cache has been corrupted, this logic can be removed.

Also harden `FileSystemLock` against other types of exceptions by not wrapping `InterruptedException` in an `IOException` and document that concurrent use on the same path is not supported.