Index branches by (line, block, branch).
A branch is marked via:
BRDA:[line_number],[block_id],[branch_id],[value]

Storing by line number requires storing multiple records per entry that
then need to be scanned through to identify the correct value to update.

This also means that the "BranchCoverage" objects now only need to be
created when writing out the combined report. Since they represent a single
line item corresponding to a single key, it's been renamed to
BranchCoverageItem.

The only real drawback here is that we must store if branches on a line
were evaluated separately. This is not a major issue though and the
simplfications elsewhere make it worthwhile.

PiperOrigin-RevId: 804285288
Change-Id: Idc61e1dfa7d5bda222faf22848204f635deccfa1