Fix parsing of text-formatted invocation policies
The --invocation_policy flag is supposed to take either a base64-encoded
representation of an InvocationPolicy proto message OR a text representation
of it.  Unfortunately, the code to handle the latter case is broken as it
is never triggered due to invalid exception handling.

Fix this by catching the correct proto and base64 decoding exception types
and falling back to text parsing in those cases.