Improve ordering of single-toolchain resolution debugging.
A single-toolchain resolution invocation produces a lot of messages when `--toolchain_resolution_debug` is set. It crucially starts with:

    INFO: ToolchainResolution: Performing resolution of <toolchain type> for target platforn <platform>

This is important to understand the context of the other messages that appear later in the sequence.

Before this change, these messages appeared first:

    INFO: ToolchainResolution: Rejected toolchain <toolchain that doesn't fit>

These are hard to read because they lack the context the `Performing resolution...` initial message provides.

This change guarantees the intro message always appears first.

Closes #26912.

Change-Id: I3ae96baa6d8d2685b51f09d2b6d6b77a902d20fa
PiperOrigin-RevId: 804907597