Use target platform's CPU constraint in configuration BEP
When `--incompatible_bep_cpu_from_platform` is enabled, the values of `Configuration.platform_name` and `Configuration.cpu` will be obtained from `--experimental_override_platform_cpu_name` mapping for the target platform or the CPU constraint of the platform if a mapping does not exist.

This is to keep the current behaviour while moving away from using `--cpu`, another change may modify that `` to represent the platform name.

PiperOrigin-RevId: 805032783
Change-Id: I2f2bd03dbd0d885e203d3dece597fb95392db243