Make action cache entries a bit more compact.
Use a single presence marker for all output metadata. A common case is to not store any.

One test needed to be updated. It was asserting on an incremental save, but that only happens when the journal is less than 1/100th of the cache size. This change made that not true.

PiperOrigin-RevId: 805431312
Change-Id: Ie471d77c61bb6e009a7f18c823d2894f6a47f140