[8.4.1] Speculative fix for repo contents cache race (#26950)
The error message observed in #26713 is consistent with
`FileChannel#open` failing because the given path doesn't exist. This
could happen if one Bazel process observed that
`!entryDir.isDirectory()`, which is true if the path doesn't exist, and
proceeded to delete the path while another process had just created the
directory and is now opening the channel. Since the entry dir is never
expected to be an existing non-directory unless the cache has been
corrupted, this logic can be removed.

Another possible source of `IOException` during normal operation is on
an interrupt (such as the user hitting Ctrl+C). Instead, follow Skyframe
best practices by surfacing this as an `InterruptedException` instead of
a `FileLockInterruptionException`.

Also document that concurrent use on the same path is not supported (it
results in an `OverlappingFileLockException` if the lock is already
held, regardless of whether that is in shared or exclusive mode) and why
the current usages are safe.

Fixes #26713

Closes #26914.

PiperOrigin-RevId: 805338728
Change-Id: Ie808ebe6113b935180b93c21679d5398aa168802

Commit
https://github.com/bazelbuild/bazel/commit/ca1cbfece457915871cc1398a667422410b4c3fc

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>