Refactor CcCompilationOutputs building in Starlark.
This change modifies how `CcCompilationOutputs` are constructed from Starlark. Instead of exposing a Java `CcCompilationOutputs.Builder` to Starlark, the Starlark code now accumulates the output artifacts in a dictionary.

A new `cc_internal.create_cc_compilation_outputs` method is introduced, which takes the contents of this dictionary as arguments and builds the `CcCompilationOutputs` object in a single call.

This reduces the number of Java methods exposed to Starlark and aligns better with Starlark's immutability principles.

PiperOrigin-RevId: 805732942
Change-Id: Ie5a60da9186e53371e95ab52562f33f6360c9d16