Migrate CcCompilationOutputs to Starlark.
This change removes the Java implementation of `CcCompilationOutputs` and replaces it with a Starlark provider `CcCompilationOutputsInfo` defined in `cc_compilation_outputs.bzl`. The Starlark implementations of `create_compilation_outputs` and `merge_compilation_outputs` are now used, and all references to the old Java-backed methods in Starlark files have been updated to use the new provider's fields and functions.

PiperOrigin-RevId: 806183545
Change-Id: If6f8579d8961cae1d07a09aedb53a58537449834