Provide a supported replacement for `Target#repr` leak of "aliasness"
`repr` of a `Target` leaked that the fact that a target is an alias as well as the alias label. This is not desirable as it allows rule implementation to introduce special handling of aliases, which would no longer make them transparent replacements for their actual targets.

Fix this by moving the alias-specific stringification to `debugPrint`, which is not accessible from Starlark. As a supported replacement that isn't exposed to Starlark rule logic, format `Target`'s as their "original label" (i.e., the `alias`' label if any) when passed to `Args`. This allows rules to emit the correct label into fixup commands.