Make actions context shareable outside of `cc_common.create_lto_backend_artifacts`
This change moves the logic for creating shareable artifacts for non-LTO backends from the Java implementation of `createLtoBackendArtifacts` to the Starlark side. The Starlark code now wraps the `actions` object with `shareable_artifacts = True` before calling the internal Java method when necessary.

The `ctx` parameter is also removed from native side.

The change is a no-op.

PiperOrigin-RevId: 806262909
Change-Id: Ic5a4d700f082e3c17d66116386e85681f23cf886