Fail if `single_version_override` version is lower than dep spec
Such a request is impossible to satisfy and likely the cause of forgetting to update the version in both places. Previously, the `bazel_dep` requirement was silently ignored.

Note that it isn't possible to entirely forbid `version` on a `bazel_dep` for an overridden module since the override would be ignored if the current module is not a root module and result in a failure in that case due to the missing version.

RELNOTES[INC]: A `single_version_override` that pins a module to a lower version than requested in a `bazel_dep` for that module now results in an error instead of silently ignoring the `bazel_dep` version requirement. This is meant to catch a common source of bugs when updating a `bazel_dep` without noticing that it is overridden.

Fixes #26964

Closes #26968.

PiperOrigin-RevId: 806383612
Change-Id: I1c283f4ccd20d6ff0bdbc57b56933819795db5e2