Fix virtual headers symlink action for C++ headers
The virtual headers action should be executed when `strip_include_prefix` attribute is not `None`, because the `.` attribute values are resolved to empty string, which is a valid value.

The virtual headers action should be independent from the original header path and include path to support library target definitions at any `BUILD` file with any project layout.

Fixes #26537

Closes #26538.

PiperOrigin-RevId: 807615188
Change-Id: I49fa14946380c0e827f046c86cdc2c33763187bd