Introduce `config.string_set`
This type for build setting rules allows the flag to accept a list of string values where only one instance of repeated values is kept and the order does not matter. This type can be used if the values order or repetition count should not affect its logical meaning and hence there is no need to fork the configuration based on these differences.

The build setting can be marked as `flag` to be used from the command line and it can be marked `repeatable` to collect values from multiple occurrences on the command line.

PiperOrigin-RevId: 807828748
Change-Id: I1d6e1a95fcf5e8b8c94bcdef5be3086e95d50e5a