zipper: handle empty filelist
updates the zipper utility so it can handle empty filelists. previously this would fail with

```
File  does not seem to exist.
```

(note the double space)

rather than being treated as an empty list (`[]`), an empty filelist was being treated as a non-empty list of length one with the only element being an empty string (`['']`)