Rollforward of https://github.com/bazelbuild/bazel/commit/d96e067331680dd4dec9fb391ed148d90daafbf8 after addressing the performance regression.
This change introduces a cache for the starlark transitions' build settings. This is because even if a transition value is evaluated and cached, the [`TransitionApplier`](https://github.com/bazelbuild/bazel/blob/master/src/main/java/com/google/devtools/build/lib/analysis/producers/TransitionApplier.java#L95-L101) computes the starlark build settings involved with the transition multiple times before checking the transition cache. This causes `StarlarkTransition.java#addLabelIfRelevant` to be more expensive specifically for the execution transition because it is called too many times checking almost all native flags to find if any of them is an alias of a starlark flag.

The logic for computing the starlark build settings of a transition is moved to `StarlarkTransitionCache` to only be evaluated if the value is not already cached.

During a single analysis, the value of `--flag_alias` should not change and hence we can cache the starlark build settings based only on the transition. After analysis phase, this cache is cleared with the transitions cache to ensure reevaluation in following builds with possibly different `--flag_alias` values.

PiperOrigin-RevId: 808250144
Change-Id: I00595d7c9e54f122d81f407a28cf6b870142fb6c