Hide `GrepIncludes` outputs from the execution graph.
The output of `GrepIncludes` is read by the include scanner and not used as an input to other spawns. Hiding it allows the execution graph to skip adding it to its `outputToNode` map (where it won't be read anyway, since it's never an input), making the `ActionInput` and its underlying exec path string eligible for GC as soon as the include scanning spawn is complete.

Note that this change does not affect the execution graph - `GrepIncludes` spawns are still included, it's just that their outputs aren't stored for use as possible inputs.

PiperOrigin-RevId: 808557649
Change-Id: I94080514803ea874bf58972890224247a7580c4b