When evaluating an artifact (directory) generated by an action template,  evaluate only the expanded action execution keys of actions that generate that directory.
The (rough) related Skyframe logic that leads to this in the ArtifactFunction is as follows:
1. Requests for the output directory
2. Gets the actions to that will generate the files in the output directory after evaluating the ActionTemplateExpansionKey for the ActionTemplate.
3. Requests for these actions to be evaluated.
4. Pieces the output files from the ActionExecutionValue(s) into the tree artifact

The previous logic requests the execution of all actions (at step 3) generated by the template, and expects that all actions generated by the action template will output a file in the output directory (at step 4).

We break this requirement by only evaluating the actions that output files within the directory (at step 3). This allows us to have 2 actions generate output files in two output directories separately in with Starlarkified action templates (instead of requiring that every action generate outputs in both output directories, which is the implication of the existing logic).

PiperOrigin-RevId: 808644280
Change-Id: I4412ab4a92b703a1e0a2ad3fbb224978799b523d